HV-1550 Clarify the build output when checkstyle fails
When checkstyle failed, it used to fail in the initial, no-log
build, and then we didn't have the checkstyle logs to tell us why it
failed. By disabling checkstyle in this initial build we can make the
build fail a little bit later, when the logs are enabled.

Also, adding a separate checkstyle execution right before the actual
build allows us to fail fast if checkstyle checks fail.