HV-1576 Let's be more generic in the definition of the 2 WildFly versions
We support 2 versions of WildFly, sometimes the current one and the next
development release, sometimes the current one and the previous one.
It's a pain to switch between the 2 so let's make it more generic.

wildfly.version is the currently supported release.
wilfly-secondary.verison is the additional version we support.